requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.100
botocore==1.34.100
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.2
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
platformdirs==4.2.2
plotly==5.22.0
ptyprocess==0.7.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-cache==1.2.0
retrying==1.3.4
rich==13.7.1
s3transfer==0.10.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starkbank-ecdsa==2.2.0
starlette==0.37.2
tenacity==8.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
url-normalize==1.4.3
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
zipp==3.18.1
